requirements.txt
streamlit>=1.24.0
opencv-python>=4.8.0
numpy>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.30.0
openai-whisper>=20231117
yt-dlp>=2023.6.22
python-dotenv>=1.0.0
scikit-learn>=1.2.0
pillow>=9.5.0
matplotlib>=3.7.0
pandas>=2.0.0
tqdm>=4.65.0
psycopg2-binary>=2.9.9
pgvector>=0.2.0
rank-bm25>=0.2.2
nltk>=3.8.1 